Add config opt for cleaning destination
This allows clients to control whether the reporter cleans the target
destination when jasmine starts.

The primary use case for this is running protractor tests in parallel
(using sharding). A new jasmine process is started for each shard, and
only the most-recent process "wins" the race to write to the report.

Disabling cleaning in this use case allows the client to clean the
target itself, and allows all of the jasmine processses to happily
append to the report.html thanks to atomic nature of fs.appendSync().